[refactor] Home.jsx 파일 components -> pages 폴더 이동
