+++ vite-project/src/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Home from './components/Home'
+import Home from './pages/Home'
 import './App.css'
 
 function App() {
