+++ vite-project/src/pages/Home.jsx
